pyproject.toml
[project]
name = "52w-stock-pdf-bot"
version = "0.1.0"
description = "Global 52w high/low stock indicator bot (US, JP, KR)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler==3.10.4",
    "fpdf2==2.7.8",
    "numpy==2.2.3",
    "pandas==2.2.3",
    "python-dotenv==1.0.1",
    "python-telegram-bot==21.1.1",
    "requests==2.32.3",
    "setuptools==75.8.0",
    "tqdm==4.67.0",
]

[tool.uv]
managed = true
